backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
# pydantic==2.5.0
python-dotenv==1.0.0
langchain==0.3.26
langchain-core==0.3.68
langchain-openai==0.3.28
openai==1.95.1
python-multipart==0.0.6
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
PyJWT==2.8.0
email-validator==2.1.0
datasets==2.14
pandas==2.1
evaluate==0.4.5
transformers==4.35.0
torch==2.1.0
ultralytics==8.0.196
opencv-python==4.8.1.78
scikit-learn==1.3.0
tqdm==4.66.1